Add MIDI CC message sending functionalities
Currently, it's possible to:
- Set a CC number to be sent (between 102 and 117)
- Toggle between two CC value modes (63/65 or 0/127)
- Send the CC message

Custom keycodes are defined in the "QK_USER" range, and can be assigned to the
encoder and keys in VIA. Refer to "midi_keycodes" in "keymap.c" for more
details.

Unfortunately, "extrakey", "mousekey", and "nkro" features had to be disabled
to leave 2(!) free endpoints for MIDI features.